backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
requests==2.31.0
python-multipart==0.0.6